# Optimize the `Trompt` example to reduce training time by ~30%